+++ CHANGELOG.md
@@ -573,3 +573,4 @@
 - [2026-04-22 23:22] perf(auth): apply RepaintBoundary on animated auth widgets
 - [2026-04-23 09:36] fix(forms): resolve race condition during forms initialization
 - [2026-04-23 09:49] perf(forms): apply RepaintBoundary on animated forms widgets
+- [2026-04-23 09:55] feat(auth): integrate custom telemetry logger in auth